LPS-179931 Move DTOConverterUtil from object-rest-impl to object-api
